FM1D2D-782: - Changed numberOfUnits from int to double. Former-commit-id: 3c96b8c90bcb5497fe8ff367c367b43665385a57 Former-commit-id: 53b15d64d0b52704bb599b906a407cf3f78af53c

+++ src/DeltaShell.Plugins.DelftModels.RainfallRunoff/Domain/Concepts/Nwrw/DryWeatherFlow.cs
@@ -33,7 +33,7 @@ namespace DeltaShell.Plugins.DelftModels.RainfallRunoff.Domain.Concepts.Nwrw
             }
         }
 
-        public virtual int NumberOfUnits { get; set; } // AVV_ENH (debiet.csv)
+        public virtual double NumberOfUnits { get; set; } // AVV_ENH (debiet.csv)
 
         
     }
